+++ transport/grpc/config.go
@@ -186,6 +186,7 @@ type transportSpec struct {
 	TransportOptions []TransportOption
 	InboundOptions   []InboundOption
 	OutboundOptions  []OutboundOption
+	DialOptions      []DialOption
 }
 
 func newTransportSpec(opts ...Option) (*transportSpec, error) {
@@ -198,6 +199,8 @@ func newTransportSpec(opts ...Option) (*transportSpec, error) {
 			transportSpec.InboundOptions = append(transportSpec.InboundOptions, opt)
 		case OutboundOption:
 			transportSpec.OutboundOptions = append(transportSpec.OutboundOptions, opt)
+		case DialOption:
+			transportSpec.DialOptions = append(transportSpec.DialOptions, opt)
 		default:
 			return nil, fmt.Errorf("unknown option of type %T: %v", o, o)
 		}
@@ -260,8 +263,7 @@ func (t *transportSpec) buildOutbound(outboundConfig *OutboundConfig, tr transpo
 		return nil, newTransportCastError(tr)
 	}
 
-	dialer := trans.NewDialer(outboundConfig.dialOptions(kit)...)
-
+	dialer := trans.NewDialer(append(outboundConfig.dialOptions(kit), t.DialOptions...)...)
 	var chooser peer.Chooser
 	if outboundConfig.Empty() {
 		if outboundConfig.Address == "" {